+++ formatter/doc.go
@@ -1,0 +1,4 @@
+// Package formatter provides functionality for formatting lint issues
+// in a human-readable format. It includes various formatters for different
+// types of issues and utility functions for text manipulation.
+package formatter
