+++ formatter/fmt_test.go
@@ -1,14 +1,19 @@
-package internal
+package formatter
 
 import (
 	"go/token"
+	"os"
+	"path/filepath"
+	"strings"
 	"testing"
 
+	"github.com/gnoswap-labs/lint/internal"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestFormatIssuesWithArrows(t *testing.T) {
-	sourceCode := &SourceCode{
+	sourceCode := &internal.SourceCode{
 		Lines: []string{
 			"package main",
 			"",
@@ -19,7 +24,7 @@ func TestFormatIssuesWithArrows(t *testing.T) {
 		},
 	}
 
-	issues := []Issue{
+	issues := []internal.Issue{
 		{
 			Rule:     "unused-variable",
 			Filename: "test.go",
@@ -55,7 +60,7 @@ error: empty-if
 	assert.Equal(t, expected, result, "Formatted output does not match expected")
 
 	// Test with tab characters
-	sourceCodeWithTabs := &SourceCode{
+	sourceCodeWithTabs := &internal.SourceCode{
 		Lines: []string{
 			"package main",
 			"",
@@ -86,7 +91,7 @@ error: empty-if
 }
 
 func TestFormatIssuesWithArrows_MultipleDigitsLineNumbers(t *testing.T) {
-	sourceCode := &SourceCode{
+	sourceCode := &internal.SourceCode{
 		Lines: []string{
 			"package main",
 			"",
@@ -101,7 +106,7 @@ func TestFormatIssuesWithArrows_MultipleDigitsLineNumbers(t *testing.T) {
 		},
 	}
 
-	issues := []Issue{
+	issues := []internal.Issue{
 		{
 			Rule:     "unused-variable",
 			Filename: "test.go",
@@ -151,7 +156,7 @@ error: example
 }
 
 func TestFormatIssuesWithArrows_UnnecessaryElse(t *testing.T) {
-	sourceCode := &SourceCode{
+	sourceCode := &internal.SourceCode{
 		Lines: []string{
 			"package main",
 			"",
@@ -165,7 +170,7 @@ func TestFormatIssuesWithArrows_UnnecessaryElse(t *testing.T) {
 		},
 	}
 
-	issues := []Issue{
+	issues := []internal.Issue{
 		{
 			Rule:     "unnecessary-else",
 			Filename: "test.go",
@@ -189,6 +194,96 @@ func TestFormatIssuesWithArrows_UnnecessaryElse(t *testing.T) {
 `
 
 	result := FormatIssuesWithArrows(issues, sourceCode)
-
 	assert.Equal(t, expected, result, "Formatted output does not match expected for unnecessary else")
+}
+
+func TestIntegratedLintEngine(t *testing.T) {
+	t.Skip("skipping integrated lint engine test")
+	tests := []struct {
+		name     string
+		code     string
+		expected []string
+	}{
+		{
+			name: "Detect unused issues",
+			code: `
+package main
+
+import (
+    "fmt"
+)
+
+func main() {
+    x := 1
+    fmt.Println("Hello")
+}
+`,
+			expected: []string{
+				"x declared and not used",
+			},
+		},
+		{
+			name: "Detect multiple issues",
+			code: `
+package main
+
+import (
+    "fmt"
+    "strings"
+)
+
+func main() {
+    x := 1
+    y := "unused"
+    fmt.Println("Hello")
+}
+`,
+			expected: []string{
+				"x declared and not used",
+				"y declared and not used",
+				`"strings" imported and not used`,
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			tmpDir, err := os.MkdirTemp("", "lint-test")
+			require.NoError(t, err)
+			defer os.RemoveAll(tmpDir)
+
+			tmpfile := filepath.Join(tmpDir, "test.go")
+			err = os.WriteFile(tmpfile, []byte(tt.code), 0o644)
+			require.NoError(t, err)
+
+			rootDir := "."
+			engine, err := internal.NewEngine(rootDir)
+			if err != nil {
+				t.Fatalf("unexpected error initializing lint engine: %v", err)
+			}
+
+			issues, err := engine.Run(tmpfile)
+			require.NoError(t, err)
+
+			assert.Equal(t, len(tt.expected), len(issues), "Number of issues doesn't match")
+
+			for _, exp := range tt.expected {
+				found := false
+				for _, issue := range issues {
+					if strings.Contains(issue.Message, exp) {
+						found = true
+						break
+					}
+				}
+				assert.True(t, found, "Expected issue not found: "+exp)
+			}
+
+			if len(issues) > 0 {
+				sourceCode, err := internal.ReadSourceCode(tmpfile)
+				require.NoError(t, err)
+				formattedIssues := FormatIssuesWithArrows(issues, sourceCode)
+				t.Logf("Found issues with arrows:\n%s", formattedIssues)
+			}
+		})
+	}
 }
